+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [Unreleased]
 ### Added
 - Syntax highlighting of fields in `let`-expressions
+- Color page in `Editor -> Color Scheme -> TableGen` 
 
 ## [0.2.0] - 2025-01-16
 ### Added